// src/app/shared/models/interfaces.ts
export type WorkOrderStatus = 'open' | 'in-progress' | 'complete' | 'blocked';

export type Timescale = 'day' | 'week' | 'month';

export interface PositionedOrder extends WorkOrderDocument {
	left: number;
	width: number;
}

export interface WorkCenterDocument {
	docId: string;
	docType: 'workCenter';
	data: {
		name: string;
	};
}

export interface WorkOrderDocument {
	docId: string;
	docType: 'workOrder';
	data: {
		name: string;
		workCenterId: string;
		status: WorkOrderStatus;
		startDate: string;
		endDate: string;
	};
}